worker/requirements.txt
pdfminer.six
pypandoc_binary
redis
requests
